Making new smaller programs for plotting buffer depth and hitmaps of cells with full buffer. WORKS! But not perfect yet

+++ BetterBkgSim/include/LayerCodeInCellID_class.h
@@ -11,8 +11,7 @@ class LayerCodeInCellID {
     virtual ~LayerCodeInCellID() {}
 
     int GetLayer(std::string const CellID, int const StartBin_layers, int const LengthBin_layers);
-    int GetLayer(int const cellid, int const StartBin_layers, int const LengthBin_layers);
-    //int GetLayer(unsigned long long const cellid, int const StartBin_layers, int const LengthBin_layers);
+    int GetLayer(long long const cellid, int const StartBin_layers, int const LengthBin_layers);
 
   private:
     //unsigned long long Layer;
