+++ BetterBkgSim/include/Subdetector_class.h
@@ -15,8 +15,7 @@ class Subdetector{
     int GetLengthLayerBin() const;				
     int GetNumberOfLayers() const;	
     int GetLayer(std::string const cellID);
-    int GetLayer(int const cellID);
-    //int GetLayer(unsigned long long const cellID);
+    int GetLayer(long long const cellID);
     int GetTotCellNumber() const;
 
 	std::vector<float> GetROOTEnergyHisto_binning() const;
